refactor: useEscapeKey 훅을 추가하여 모달에서 Escape 키 처리 로직 분리
